% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjacency_mat.R
\name{get_adjacency_mat}
\alias{get_adjacency_mat}
\title{Get adjacency matrix}
\usage{
get_adjacency_mat(data, k = 20)
}
\arguments{
\item{data}{a matrix whose rows represent samples and columns represent features}

\item{k}{how many k nearest neighbors to build adjacency matrix}
}
\value{
Returns a adjacency matrix
}
\description{
Get adjacency matrix
}
