% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cutoff.R
\name{get_cutoff}
\alias{get_cutoff}
\title{get cutoff based on the best F1 score}
\usage{
get_cutoff(value, labels, true_label = "TRUE", plot = F)
}
\arguments{
\item{plot}{}
}
\value{
Return the best cutoff
}
\description{
get cutoff based on the best F1 score
}
