% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ROC.R
\name{get_cutoffs}
\alias{get_cutoffs}
\title{get_cutoffs}
\usage{
get_cutoffs(
  true,
  pred_score,
  expected_precision0 = 0.9,
  expected_precision1 = 0.9,
  plot = F
)
}
\arguments{
\item{plot}{}
}
\description{
get_cutoffs
}
