% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rank.R
\name{get_quantile}
\alias{get_quantile}
\title{Get quantile of a number in a vector}
\usage{
get_quantile(data, x)
}
\arguments{
\item{data}{a vector}

\item{x}{a number}
}
\value{
the quantile of the number
}
\description{
Get quantile of a number in a vector
}
