% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rank.R
\name{get_rank}
\alias{get_rank}
\title{Get ranks of a series of numerics with names}
\usage{
get_rank(x)
}
\arguments{
\item{x}{a series of numerics with names}
}
\value{
each element's rank
}
\description{
Get ranks of a series of numerics with names
}
