% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_RF.R
\name{my_RF}
\alias{my_RF}
\title{Random forest code memo}
\usage{
my_RF(train, labels, test, verbose = T)
}
\arguments{
\item{train}{The train dataset whose rows are samples and columns are features}

\item{labels}{Labels of the train dataset}

\item{test}{The test dataset whose rows are samples and columns are features}

\item{verbose}{Whether to display messages}
}
\value{
todo
}
\description{
Random forest code memo
}
