% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo_install_packages.R
\name{myinstall_with_C99}
\alias{myinstall_with_C99}
\title{Install R packages with C99 standard}
\usage{
myinstall_with_C99(package.name, via = "CRAN")
}
\arguments{
\item{package.name}{R package name(s)}

\item{via}{'CRAN' or 'bioc'(bioconductor)}
}
\value{

}
\description{
Install R packages with C99 standard
}
