% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mykmeans.R
\name{mykmeans}
\alias{mykmeans}
\title{Perform kmeans with multiple 'k's and automatically select the best k with the}
\usage{
mykmeans(x, k)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a matrix (such as a numeric vector or a data frame with all numeric columns).}

\item{k}{a set of candidate centers}
}
\value{
returns an object of class "kmeans".
}
\description{
Perform kmeans with multiple 'k's and automatically select the best k with the maximum silhouette
}
