% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myknn.R
\name{myknn}
\alias{myknn}
\title{Several K Nearest Neighbors (KNN) methods}
\usage{
myknn(
  train,
  labels,
  test,
  k,
  method = "weighted",
  dist.metric = "cosine",
  verbose = T
)
}
\arguments{
\item{train}{The train dataset whose rows are samples and columns are features}

\item{labels}{Labels of the train dataset}

\item{test}{The test dataset whose rows are samples and columns are features}

\item{k}{Number of neighbors}

\item{method}{Knn method; can be one of "weighted", "Shieh", "normal","self_weighted"
 \itemize{
  \item{'weighted': }{weighted knn}
  \item{'Shieh': }{https://ieeexplore.ieee.org/document/6298627}
  \item{'normal': }{normal knn}
}}

\item{dist.metric}{Distance metric; can be one of "euclidean", "cosine", "manhattan",}

\item{verbose}{Whether to display messages}
}
\value{
Return a data frame containing a predicted labels column and a score column.
}
\description{
Several K Nearest Neighbors (KNN) methods
}
