% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot_jupyter_settings.R
\name{myplot_jupyter_settings}
\alias{myplot_jupyter_settings}
\title{my_plot_jupyter_settings}
\usage{
myplot_jupyter_settings(width = 10, height = 10)
}
\arguments{
\item{height}{}
}
\value{
todo
}
\description{
my_plot_jupyter_settings
}
