% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot_ggplot.R
\name{myplot_scatter_with_density}
\alias{myplot_scatter_with_density}
\title{Scatter plot with density plot of x and y axis}
\usage{
myplot_scatter_with_density()
}
\description{
Scatter plot with density plot of x and y axis
}
