% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ROC.R
\name{plot_PRC}
\alias{plot_PRC}
\title{plot P-R Curve}
\usage{
plot_PRC(true, pred_prob, length.out = 50, return_df = F)
}
\arguments{
\item{return_df}{}
}
\value{
either a plot (default) or a result dataframe(return_df = T)
}
\description{
plot P-R Curve
}
