% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_h5ad_as_Seurat.R
\name{read_h5ad_as_Seurat}
\alias{read_h5ad_as_Seurat}
\title{read_h5ad_as_Seurat}
\usage{
read_h5ad_as_Seurat(h5ad_file_name)
}
\arguments{
\item{h5ad_file_name}{}
}
\description{
read_h5ad_as_Seurat
}
