% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRF.R
\name{RF_CV}
\alias{RF_CV}
\title{Random forest cross validation}
\usage{
RF_CV(trainx, trainy, n.fold = 5, seed = 123)
}
\arguments{
\item{trainx}{trainx}

\item{trainy}{trainy}

\item{n.fold}{n.fold}

\item{seed}{seed}
}
\value{
dataframe of true and pred_prob
}
\description{
Random forest cross validation
}
