% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{volcano_plot}
\usage{
volcano_plot(de_genes, log2FC = 0.2, padj = 0.05, highlight_num = 10)
}
\arguments{
\item{highlight_num}{}
}
\description{
volcano_plot
}
