% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cutoff.R
\name{get_cutoff}
\alias{get_cutoff}
\title{Get cutoff based on the best F1 score}
\usage{
get_cutoff(
  value,
  labels,
  true_label = "TRUE",
  min.F1 = NULL,
  alternative_cutoffs = NULL,
  plot = F
)
}
\arguments{
\item{value}{a vector of values}

\item{labels}{the labels of a vector of values}

\item{true_label}{true labels}

\item{min.F1}{the minimum F1}

\item{alternative_cutoffs}{alternative cutoffs}

\item{plot}{whether to plot}
}
\value{
Return the best cutoff
}
\description{
Get cutoff based on the best F1 score
}
