% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mykmeans.R
\name{mykmeans}
\alias{mykmeans}
\title{Perform kmeans with multiple 'k's and automatically select the best k}
\usage{
mykmeans(x, k, iter.max = 10)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a matrix.}

\item{k}{a set of candidate centers.}

\item{iter.max}{the maximum number of iterations allowed.}
}
\value{
returns an object of class "kmeans".
}
\description{
Perform kmeans with multiple 'k's and automatically select the best k with the maximum average silhouette
}
